pyproject.toml
[tool.poetry]
authors = [
  "Laureηt <laurentfainsin@protonmail.com>",
  "Damien Guillotin <damien.guillotin@etu.inp-n7.fr>",
  "Pierre-Eliot Jourdan <pierreliot.jourdan@etu.inp-n7.fr>",
]
description = "Rocket League sideswipe image classification deep learning project"
license = "MIT"
name = "DeepSwipe"
version = "0.1.0"

[tool.poetry.dependencies]
Pillow = "^9.0.1"
numpy = "^1.22.3"
pure-python-adb = "^0.3.0-alpha.0"
python = ">=3.10,<3.11"
tensorflow = "^2.8.0"

[tool.poetry.dev-dependencies]
black = "^22.1"
isort = "^5.9.3"
mypy = "^0.931"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
exclude = '''
/(
  \.git
  \.venv
)/
'''
include = '\.pyi?$'
line-length = 120
target-version = ["py310"]

[tool.isort]
multi_line_output = 3
profile = "black"
